refactor: Standardize command package structure and module configuration
- Introduce module.go files for httpd, job, and search commands
- Simplify command package organization with consistent naming
- Remove init() functions and replace with Command() method
- Streamline Fx module dependencies for each command
- Improve package-level documentation for clarity
- Enhance command registration and configuration handling
